Add files via upload
Removal of redundant code for the word model. 
Now, you can simply train a line model which will work for both line and word.

Improving the code structure for the Model file. This file is now divided into Model_Testing and Model_Training which contain the code and model structure for testing and training respectively.